// frontend/src/api/models/create-counseling-plan-phase-goal-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateCounselingPlanPhaseGoalDTO
 */
export interface CreateCounselingPlanPhaseGoalDTO {
    /**
     * 
     * @type {string}
     * @memberof CreateCounselingPlanPhaseGoalDTO
     */
    'counselingPlanPhaseId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateCounselingPlanPhaseGoalDTO
     */
    'goalName'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateCounselingPlanPhaseGoalDTO
     */
    'goalDescription'?: string;
}